Add support for zeus console
When a `.zeus.sock` file is present in the current directory this means
that the project uses [zeus](https://github.com/burke/zeus) and would be
suited to use it to launch a console. Depending on if zeus is installed
globally or for the current project only this might require `bundle
exec` to run.