Improve inf-ruby-console-rails-p
Remove the requirements depending on the existance and location of a
Gemfile.lock file, as these prevent the proper operation of inf-ruby
when the Gemfile.lock is elsewhere, or Bundler is not in use.

Instead, use the existance of bin/rails, as this is something that the
Rails tooling itself depends on, and therefore more likely to be both
indicative of Rails being used, and required, as Rails is difficult to
use without it.